Added more CRUD DB helpers

+++ SampleCode-V5/android-sdk-v5-sample/src/main/java/za/co/lsmc/data/HeadFrame.kt
@@ -1,4 +1,4 @@
 package za.co.lsmc.data
 
-data class HeadFrame(var id: Long, var siteId: Int, var altitudeM: Double) {
+data class HeadFrame(var id: Long, var siteId: Long, var altitudeM: Double) {
 }
